Using System.Xml for build html file, also tiny edits.

+++ TestApp/Model/FileInformation.cs
@@ -21,7 +21,7 @@ namespace TestApp
         /// <summary>
         /// Размер файла в байтах
         /// </summary>
-        public float FileSize { get; set; }
+        public long FileSize { get; set; }
 
         /// <summary>
         /// Абсолютный путь вместе с названием файла
